We need to add variables for the test message.
We were not given the information about how the text message should be formatted.  We were told that there should be a specific format for the text message so that the server could read it.
We were not given the number that text message will need to be sent to.  We were told that there is a different number that the text message needs to be sent to depending on where the message is sent from.
If we don't know what it should look like, we should at least make the app send some kind of text message.